+++ tests/ttnn/unit_tests/operations/fused/test_softmax.py
@@ -8,7 +8,7 @@ import torch
 import torch.nn.functional as F
 
 import ttnn
-from tests.ttnn.utils_for_testing import assert_numeric_metrics, assert_with_pcc
+from tests.ttnn.utils_for_testing import assert_numeric_metrics
 from models.common.utility_functions import torch_random
 
 TEST_PADDING_VALUE = -42
@@ -18,6 +18,7 @@ TEST_PADDING_VALUE = -42
 @pytest.mark.parametrize(
     "batch_size, h, w, dim",
     [
+        (1, 128, 128000, -1),
         (1, 32, 128000, -1),
         (1, 2048, 32000, -1),
         (1, 512, 32000, -1),
@@ -771,55 +772,3 @@ def test_softmax_large_kernel_mask_padded(device, shape, dim):
         ulp_threshold=15,
         check_ulp=True,
     )
-
-
-@pytest.mark.parametrize("target_sequence_size", [8192])
-@pytest.mark.parametrize("sequence_size", [384])
-@pytest.mark.parametrize("iterations", [50])
-def test_transformer_attention_softmax_inplace_large_kernel_stress(
-    device, target_sequence_size, sequence_size, iterations
-):
-    """Regression test for issue #45200: in-place attention_softmax_ on a tensor large enough to
-    select the large-tensor kernel used to be blocked by a "Tensor is too large to run softmax
-    inplace" assertion, and removing that assertion hung. The real cause was the fused scale+mask
-    large kernel passing the full capped CB length instead of the per-chunk length, which
-    over-consumed the input CB (and deadlocked dispatch) whenever Wt was not a multiple of the
-    capped CB length. Verify the op now runs in-place without hanging and returns deterministic,
-    correct results across repeated iterations.
-
-    Note: the input range is kept moderate because attention_softmax_ uses numeric_stable=False, so
-    extreme magnitudes would overflow the (unstable) exp and diverge from the stable torch golden -
-    a numerics artifact unrelated to this fix (see issue #28525)."""
-    torch.manual_seed(0)
-
-    input_shape = (1, 1, sequence_size, target_sequence_size)
-    torch_input_tensor = torch_random(input_shape, -5.0, 5.0, dtype=torch.bfloat16)
-    torch_attention_mask = torch_random(input_shape, 0, 1.0, dtype=torch.bfloat16)
-
-    golden_function = ttnn.get_golden_function(ttnn.transformer.attention_softmax_)
-    torch_output_tensor = golden_function(torch_input_tensor, head_size=None, attention_mask=torch_attention_mask)
-
-    # Mask is read-only across iterations, so upload once.
-    attention_mask = ttnn.from_torch(torch_attention_mask, device=device, dtype=ttnn.bfloat16, layout=ttnn.TILE_LAYOUT)
-
-    first_output = None
-    for i in range(iterations):
-        # Re-upload the input every iteration because attention_softmax_ mutates it in place.
-        input_tensor = ttnn.from_torch(torch_input_tensor, device=device, dtype=ttnn.bfloat16, layout=ttnn.TILE_LAYOUT)
-        input_addr = input_tensor.buffer_address()
-        output_tensor = ttnn.transformer.attention_softmax_(
-            input_tensor, head_size=None, attention_mask=attention_mask, causal_mask=True
-        )
-        # The op must remain truly in-place (output aliases the input buffer).
-        assert output_tensor.buffer_address() == input_addr
-        output_tensor = ttnn.to_torch(ttnn.from_device(output_tensor))
-        assert_with_pcc(torch_output_tensor, output_tensor, 0.99)
-
-        if first_output is None:
-            first_output = output_tensor
-        else:
-            assert torch.equal(first_output, output_tensor), (
-                f"In-place large-kernel softmax is non-deterministic at iteration {i}: "
-                f"output differs from iteration 0 (max abs delta "
-                f"{(first_output - output_tensor).abs().max().item()})"
-            )
